[finishes #166259565] Add user login endpoint
