+++ package.json
@@ -5,11 +5,11 @@
     "plugin:jasmine"
   ],
   "description": "An online market place for automobiles where  users can sell their cars or buy from trusted dealerships or private sellers.",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
-    "start": "nodemon --exec babel-node index.js",
-    "test-init": "jasmine --init",
-    "test": "nyc jasmine",
+    "start": "nodemon --exec babel-node src/index.js",
+    "test-init": "jasmine init",
+    "test": "nyc jasmine --config=src/spec/support/jasmine.json",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
